+++ wedpr-components/user/src/main/java/com/webank/wedpr/components/user/response/GetGroupCountResponse.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GetGroupCountResponse {
-    private Integer groupCount;
+    private Long groupCount;
 }
